pyproject.toml
[project]
name = "agent-rag-mcp"
version = "0.1.0"
description = "Dynamic RAG MCP server for AI agents - reduces hallucinations and repetitive errors"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.14.1",
    "google-genai>=1.56.0",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "toon-format>=0.1.0",
    "weaviate-client>=4.19.0",
]

[project.scripts]
agent-rag-mcp = "agent_rag_mcp.server.main:main"
agent-rag-client = "agent_rag_mcp.client.main:main"

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_rag_mcp"]
